Bow works, sheild works, and reset works.

+++ classes.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 
+
 @dataclass
 class Rect:
     x: int
@@ -8,18 +9,22 @@ class Rect:
     h: int
     color: int
 
+
 @dataclass
 class Agent:
     x: int
     y: int
 
+
 @dataclass
 class Enemy(Agent):
     health: int
 
+
 @dataclass
 class Moblin(Enemy):
-   pass 
+    pass
+
 
 @dataclass
 class Item:
@@ -27,14 +32,15 @@ class Item:
     y: int
     name: str
     tile_x: int
-    tile_y: int  
- 
+    tile_y: int
+
+
 @dataclass
 class ItemWithDirection:
     x: int
     y: int
     tile_x_down: int
-    tile_y_down: int      
+    tile_y_down: int
     tile_x_up: int
     tile_y_up: int
     tile_x_left: int
@@ -42,7 +48,8 @@ class ItemWithDirection:
     tile_x_right: int
     tile_y_right: int
     alpha: int
- 
+
+
 @dataclass
 class Player(Agent):
     inventory: list[Item]
@@ -52,7 +59,7 @@ class Player(Agent):
     shielding: bool
     arrow_frame: int
     arrow_dir: str
-    health: int
+    health: float
 
 
 @dataclass
@@ -61,6 +68,7 @@ class World:
     win: bool
     player: Player
     walls: list[Rect]
+
 
 # world = World(
 #     game_over = True,
